pyproject.toml
[project]
name = "godot-love-wechat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=2.9.0",
    "pywebview>=5.3.2",
]
